+++ src/main/java/com/starkbank/BoletoPayment.java
@@ -14,16 +14,16 @@ import java.util.Map;
 public final class BoletoPayment extends Resource {
     static ClassData data = new ClassData(BoletoPayment.class, "BoletoPayment");
 
-    public final Long amount;
-    public final String taxId;
-    public final String description;
-    public final String line;
-    public final String barCode;
-    public final String scheduled;
-    public final String[] tags;
-    public final String status;
-    public final Integer fee;
-    public final String created;
+    public Long amount;
+    public String taxId;
+    public String description;
+    public String line;
+    public String barCode;
+    public String scheduled;
+    public String[] tags;
+    public String status;
+    public Integer fee;
+    public String created;
 
     /**
      * BoletoPayment object