+++ src/main/java/com/starkbank/utils/HttpClient.java
@@ -6,6 +6,7 @@ import retrofit2.Retrofit;
 
 import java.util.concurrent.TimeUnit;
 
+
 final class HttpClient {
     private static ClientService productionInstance;
     private static ClientService sandboxInstance;