+++ src/test/java/TestBoletoPayment.java
@@ -21,20 +21,8 @@ public class TestBoletoPayment {
     public void testCreate() throws Exception {
         Settings.user = utils.User.defaultProject();
         List<BoletoPayment> payments = new ArrayList<>();
-        HashMap<String, Object> data = new HashMap<>();
-        int randomNum = ThreadLocalRandom.current().nextInt(1, 100000000);
 
-        LocalDateTime tomorrow = LocalDateTime.now().plusDays(1);
-        DateTimeFormatter dateFormat = DateTimeFormatter.ofPattern("yyyy-MM-dd", Locale.ENGLISH);
-        String boletoDay = dateFormat.format(tomorrow);
-        
-
-        data.put("line", "34191.09107 05447.947309 71444.640008 8 846600" + String.format("%08d", randomNum));
-        data.put("taxId", "38.435.677/0001-25");
-        data.put("scheduled", boletoDay);
-        data.put("description", "Payment for killing white walkers");
-        data.put("tags", new String[]{"little girl", "no one"});
-        payments.add(new BoletoPayment(data));
+        payments.add(TestBoletoPayment.example(true));
 
         payments = BoletoPayment.create(payments);
 
@@ -75,19 +63,8 @@ public class TestBoletoPayment {
     public void testCreateAndDelete() throws Exception {
         Settings.user = utils.User.defaultProject();
         List<BoletoPayment> payments = new ArrayList<>();
-        HashMap<String, Object> data = new HashMap<>();
-        int randomNum = ThreadLocalRandom.current().nextInt(1, 100000000);
 
-        LocalDateTime tomorrow = LocalDateTime.now().plusDays(1);
-        DateTimeFormatter dateFormat = DateTimeFormatter.ofPattern("yyyy-MM-dd", Locale.ENGLISH);
-        String boletoDay = dateFormat.format(tomorrow);
-
-        data.put("line", "34191.09107 05447.947309 71444.640008 8 846600" + String.format("%08d", randomNum));
-        data.put("taxId", "38.435.677/0001-25");
-        data.put("scheduled", boletoDay);
-        data.put("description", "Payment for killing white walkers");
-        data.put("tags", new String[]{"little girl", "no one"});
-        payments.add(new BoletoPayment(data));
+        payments.add(TestBoletoPayment.example(true));
 
         payments = BoletoPayment.create(payments);
 
@@ -116,5 +93,24 @@ public class TestBoletoPayment {
             System.out.println(log);
         }
         Assert.assertTrue(i > 0);
+    }
+
+    static BoletoPayment example(boolean scheduled) throws Exception{
+        HashMap<String, Object> data = new HashMap<>();
+        int randomNum = ThreadLocalRandom.current().nextInt(1, 100000000);
+
+        data.put("line", "34191.09107 05447.947309 71444.640008 8 846600" + String.format("%08d", randomNum));
+        data.put("taxId", "38.435.677/0001-25");
+        data.put("description", "Payment for killing white walkers");
+        data.put("tags", new String[]{"little girl", "no one"});
+
+        if(scheduled) {
+            LocalDateTime tomorrow = LocalDateTime.now().plusDays(1);
+            DateTimeFormatter dateFormat = DateTimeFormatter.ofPattern("yyyy-MM-dd", Locale.ENGLISH);
+            String boletoDay = dateFormat.format(tomorrow);
+            data.put("scheduled", boletoDay);
+        }
+
+        return new BoletoPayment(data);
     }
 }